+++ libs/components/forms/src/lib/modules/checkbox/fixtures/standard-checkbox-group.component.ts
@@ -24,8 +24,10 @@ export class SkyStandardCheckboxGroupComponent {
   protected formGroup: FormGroup;
   protected contactMethod: FormGroup;
 
+  public helpKey: string | undefined;
+  public helpPopoverContent: string | undefined;
   public hintText: string | undefined;
-  public labelHidden = false;
+  public headingHidden = false;
   public required = false;
   public stacked = true;