+++ libs/components/forms/testing/src/checkbox/checkbox-group-harness.spec.ts
@@ -1,5 +1,7 @@
 import { TestbedHarnessEnvironment } from '@angular/cdk/testing/testbed';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { SkyHelpService } from '@skyux/core';
+import { SkyHelpTestingModule } from '@skyux/core/testing';
 
 import { SkyCheckboxGroupHarness } from './checkbox-group-harness';
 import { CheckboxHarnessTestComponent } from './fixtures/checkbox-harness-test.component';
@@ -10,7 +12,7 @@ async function setupTest(options: { dataSkyId?: string } = {}): Promise<{
   fixture: ComponentFixture<CheckboxHarnessTestComponent>;
 }> {
   await TestBed.configureTestingModule({
-    imports: [CheckboxHarnessTestModule],
+    imports: [CheckboxHarnessTestModule, SkyHelpTestingModule],
   }).compileComponents();
 
   const fixture = TestBed.createComponent(CheckboxHarnessTestComponent);
@@ -28,42 +30,42 @@ async function setupTest(options: { dataSkyId?: string } = {}): Promise<{
 }
 
 describe('Checkbox group harness', () => {
-  it('should get the label text', async () => {
+  it('should get the heading text', async () => {
     const { checkboxGroupHarness } = await setupTest();
 
-    await expectAsync(checkboxGroupHarness.getLabelText()).toBeResolvedTo(
+    await expectAsync(checkboxGroupHarness.getHeadingText()).toBeResolvedTo(
       'Contact method',
     );
   });
 
-  it('should get the label text when label text is hidden', async () => {
+  it('should get the heading text when heading text is hidden', async () => {
     const { checkboxGroupHarness, fixture } = await setupTest({
       dataSkyId: 'checkbox-group',
     });
 
-    fixture.componentInstance.hideGroupLabel = true;
+    fixture.componentInstance.hideGroupHeading = true;
     fixture.detectChanges();
 
-    await expectAsync(checkboxGroupHarness.getLabelText()).toBeResolvedTo(
+    await expectAsync(checkboxGroupHarness.getHeadingText()).toBeResolvedTo(
       'Contact method',
     );
   });
 
-  it('should indicate the label is not hidden', async () => {
+  it('should indicate the heading is not hidden', async () => {
     const { checkboxGroupHarness } = await setupTest();
 
-    await expectAsync(checkboxGroupHarness.getLabelHidden()).toBeResolvedTo(
+    await expectAsync(checkboxGroupHarness.getHeadingHidden()).toBeResolvedTo(
       false,
     );
   });
 
-  it('should indicate the label is hidden', async () => {
+  it('should indicate the heading is hidden', async () => {
     const { checkboxGroupHarness, fixture } = await setupTest();
 
-    fixture.componentInstance.hideGroupLabel = true;
+    fixture.componentInstance.hideGroupHeading = true;
     fixture.detectChanges();
 
-    await expectAsync(checkboxGroupHarness.getLabelHidden()).toBeResolvedTo(
+    await expectAsync(checkboxGroupHarness.getHeadingHidden()).toBeResolvedTo(
       true,
     );
   });
@@ -113,18 +115,35 @@ describe('Checkbox group harness', () => {
     ).toBeRejectedWithError('No help inline found.');
   });
 
-  it('should open help inline popover', async () => {
+  it('should open help inline popover when clicked', async () => {
     const { checkboxGroupHarness, fixture } = await setupTest();
+
     fixture.componentInstance.helpPopoverContent = 'popover content';
     fixture.detectChanges();
 
     await checkboxGroupHarness.clickHelpInline();
     fixture.detectChanges();
-    fixture.whenStable();
+    await fixture.whenStable();
 
     await expectAsync(
       checkboxGroupHarness.getHelpPopoverContent(),
     ).toBeResolved();
+  });
+
+  it('should open global help widget when clicked', async () => {
+    const { checkboxGroupHarness, fixture } = await setupTest();
+    const helpSvc = TestBed.inject(SkyHelpService);
+    const helpSpy = spyOn(helpSvc, 'openHelp');
+
+    fixture.componentInstance.helpPopoverContent = undefined;
+    fixture.componentInstance.helpKey = 'helpKey.html';
+    fixture.detectChanges();
+
+    await checkboxGroupHarness.clickHelpInline();
+    fixture.detectChanges();
+    await fixture.whenStable();
+
+    expect(helpSpy).toHaveBeenCalledWith({ helpKey: 'helpKey.html' });
   });
 
   it('should get help popover content', async () => {
@@ -134,7 +153,7 @@ describe('Checkbox group harness', () => {
 
     await checkboxGroupHarness.clickHelpInline();
     fixture.detectChanges();
-    fixture.whenStable();
+    await fixture.whenStable();
 
     await expectAsync(
       checkboxGroupHarness.getHelpPopoverContent(),
@@ -149,7 +168,7 @@ describe('Checkbox group harness', () => {
 
     await checkboxGroupHarness.clickHelpInline();
     fixture.detectChanges();
-    fixture.whenStable();
+    await fixture.whenStable();
 
     await expectAsync(
       checkboxGroupHarness.getHelpPopoverTitle(),