+++ libs/components/help-inline/testing/src/help-inline/help-inline-harness.ts
@@ -41,7 +41,7 @@ export class SkyHelpInlineHarness extends SkyComponentHarness {
       );
     }
 
-    await button.click();
+    return button.click();
   }
 
   /**